+++ app/Http/Services/CategoryService.php
@@ -55,6 +55,10 @@ class CategoryService extends BaseController{
 
     public function delete($request)
     {
+        $id = $this->category->find($request['id']);
+        if (!$id) {
+            return $this->sendError("No Record to Delete");
+        }
         $this->deleteById($request['id'], 'categories');
         return $this->sendResponse('Category Delete Success');
     }
