+++ app/Http/Services/KitchenService.php
@@ -72,6 +72,10 @@ class KitchenService extends BaseController
 
     public function delete($request)
     {
+        $id = $this->kitchen->find($request['id']);
+        if (!$id) {
+            return $this->sendError("No Record to Delete");
+        }
         $this->deleteById($request['id'],'kitchens');
         return $this->sendResponse('Kitchen Delete Success');
     }
