+++ app/Http/Services/OperatorService.php
@@ -78,6 +78,10 @@ class OperatorService extends BaseController{
 
     public function delete($request)
     {
+        $id = $this->operator->find($request['id']);
+        if (!$id) {
+            return $this->sendError("No Record to Delete");
+        }
         $this->deleteById($request['id'],'operators');
         return $this->sendResponse('Operator Delete Success');
     }
